# apps/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: ProSlides API
  version: 0.1.0
  description: Contract-first API for the Go modular monolith.
servers:
  - url: /
paths:
...
  /api/v1/auth/register:
    post:
      operationId: registerAccount
      summary: Create an account and either start a session or begin email verification.
      description: When email verification is disabled, sets an HttpOnly session cookie and a readable CSRF cookie. Otherwise no session is issued until the six-digit code is verified. Passwords and verification codes are never returned or logged; OTPs are stored only as keyed hashes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created. The response states whether verification is required and whether a session was established.
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationResult'
        '400': { $ref: '#/components/responses/ValidationError' }
        '409':
          description: An account already uses this email address.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { description: Verification delivery is required but unavailable. }
  /api/v1/auth/login:
    post:
      operationId: loginAccount
      summary: Start an authenticated session.
      description: Sets an HttpOnly session cookie and a readable CSRF cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Session established.
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentUser' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { description: Credentials are valid but email verification is still required. }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/auth/logout:
    post:
      operationId: logoutAccount
      summary: End the current authenticated session.
      description: Requires the session cookie and `X-CSRF-Token` equal to the CSRF cookie value.
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204': { description: Session ended, whether or not it had already expired. }
        '403': { $ref: '#/components/responses/CSRFError' }
  /api/v1/auth/me:
    get:
      operationId: getCurrentUser
      summary: Return the account represented by the current session cookie.
      responses:
        '200':
          description: Current authenticated account.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentUser' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/verify:
    post:
      operationId: verifyAccountEmail
      summary: Activate an account with its one-time six-digit email code.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/VerifyEmailRequest' } } }
      responses:
        '200':
          description: Email verified and the normal session and CSRF cookies issued.
          content: { application/json: { schema: { $ref: '#/components/schemas/CurrentUser' } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/auth/verify/resend:
    post:
      operationId: resendAccountVerification
      summary: Replace and resend an email verification code without disclosing unknown accounts.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ResendVerificationRequest' } } }
      responses:
        '202':
          description: Accepted whether or not the account exists.
          content: { application/json: { schema: { $ref: '#/components/schemas/VerificationDeliveryResult' } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { description: Verification delivery is temporarily unavailable. }
  /api/v1/auth/google:
    post:
      operationId: authenticateWithGoogle
      summary: Verify a signed Google ID token and issue the normal session and CSRF cookies.
      description: The backend validates the RS256 signature against Google's JWKS, issuer, audience, expiry, and verified-email claim. Unsigned claim decoding is prohibited.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/GoogleAuthRequest' } } }
      responses:
        '200':
          description: Google identity accepted and session established.
          content: { application/json: { schema: { $ref: '#/components/schemas/GoogleAuthResult' } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { description: Google authentication is not configured or its public keys are unavailable. }
  /api/v1/auth/password/reset:
    post:
      operationId: requestPasswordReset
      summary: Send a one-time password reset link without disclosing account existence.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PasswordResetRequest' } } }
      responses:
        '202': { description: Request accepted whether or not the email exists. }
        '400': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { description: Password reset delivery is temporarily unavailable. }
  /api/v1/auth/password/reset/confirm:
    post:
      operationId: confirmPasswordReset
      summary: Consume a one-time reset token and revoke all existing account sessions.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PasswordResetConfirmRequest' } } }
      responses:
        '204': { description: Password changed and existing sessions revoked. }
        '400': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
...
  schemas:
    RegisterRequest:
      type: object
      required: [email, display_name, password]
      properties:
        email: { type: string, format: email, maxLength: 320 }
        display_name: { type: string, minLength: 1, maxLength: 100 }
        password: { type: string, format: password, minLength: 12, maxLength: 128 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, format: password, maxLength: 128 }
    VerifyEmailRequest:
      type: object
      required: [email, code]
      properties:
        email: { type: string, format: email, maxLength: 320 }
        code: { type: string, pattern: '^[0-9]{6}$' }
    ResendVerificationRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email, maxLength: 320 }
    VerificationDeliveryResult:
      type: object
      required: [detail, resend_seconds, code_expires_in_seconds]
      properties:
        detail: { type: string }
        resend_seconds: { type: integer, minimum: 0 }
        code_expires_in_seconds: { type: integer, minimum: 0 }
    GoogleAuthRequest:
      type: object
      required: [token]
      properties:
        token: { type: string, minLength: 64, maxLength: 16384 }
    GoogleAuthResult:
      allOf:
        - { $ref: '#/components/schemas/CurrentUser' }
        - type: object
          required: [needs_password_setup, is_new_user]
          properties:
            name: { type: string }
            full_name: { type: string }
            needs_password_setup: { type: boolean }
            is_new_user: { type: boolean }
    PasswordResetRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email, maxLength: 320 }
    PasswordResetConfirmRequest:
      type: object
      required: [uid, token, new_password]
      properties:
        uid: { type: string, format: uuid }
        token: { type: string, minLength: 32, maxLength: 128 }
        new_password: { type: string, format: password, minLength: 12, maxLength: 128 }
    RegistrationResult:
      allOf:
        - { $ref: '#/components/schemas/CurrentUser' }
        - type: object
          required: [verification_sent]
          properties:
            verification_sent: { type: boolean }
            resend_seconds: { type: integer, minimum: 0 }
            code_expires_in_seconds: { type: integer, minimum: 0 }
    CurrentUser:
      type: object
      required: [id, email, display_name, is_active]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string }
        is_active: { type: boolean }
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
  responses:
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: No valid current session exists.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    CSRFError:
      description: The CSRF header and cookie do not match a valid current session.
      content:
        application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
    RateLimited:
      description: The bounded authentication request budget was exceeded.
      headers:
        Retry-After: { schema: { type: integer, minimum: 1 } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }